Fix compatibility with PyPy
When trying to import plotext in a script running under PyPy, the following stack trace will be seen:

```
~/$ pyenv/bin/python -c "import plotext as plt; pizzas = ['Sausage', 'Pepperoni', 'Mushrooms', 'Cheese', 'Chicken', 'Beef']; percentages = [14, 36, 11, 8, 7, 4]; plt.simple_bar(pizzas, percentages, width = 100, title = 'Most Favored Pizzas in the World'); plt.show()"
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/robert/pyenv/site-packages/plotext/__init__.py", line 6, in <module>
    from ._core import *
  File "/home/robert/pyenv/site-packages/plotext/_core.py", line 12, in <module>
    import plotext._global as _glob
  File "/home/robert/pyenv/site-packages/plotext/_global.py", line 7, in <module>
    import plotext._global as glob
AttributeError: module 'plotext' has no attribute '_global'
~/$
```

I don't see a great reason to have the module try to import itself and it doesn't seem to work on PyPy (although it does work in cPython). Removing the self-import and fixing the one place that uses it to just reference the already global symbol fixes the issue.